- upgraded packages, causing a mess. i had wanted to upgrade python   interpreter for some reason -- mine was on 3.10, i upgraded to 3.12,   lot's of packages wouldn't work even after trying to upgrade them all,   so went back to 3.10, but had problems reverting all the package   upgrades. ended up with something that mostly worked, except:   - json serialization of Row and RowMapping is broken (again, I can't     remember how we fixed it way back when). I fixed it in sql_query by     converting to list and dict, but several other functions will need     to be fixed. @joeflack4 -- sorry about all this. let's talk about it     monday.   - also needed explicit type cast on concepts_with_counts.total_cnt     because of json serialization problems - fixed incorrect return types on several route handlers - was using a networkx pickle function for creating graph files; it   wouldn't work after upgrade to networkx 4. ended up having to   downgrade to 3.2.1 for some other reason, but leaving the code in   place to use regular pickle instead since it works either way   - the pickle files are much bigger now. i don't know if that's because     of more data than when we made them before or because of differences     in the pickling algorithm. will commit and push to termhub-vocab to     update them for everyone - python@3.12 gives error on backslash escaped characters in regular   strings; had to make them r'...' strings; no reason to undo that in   going back to 3.10 - got rid of caching and mezmorize in medication_csets_from_rxnorm_to_pg   because of package problems. we're not using that file right now   anyway

+++ backend/db/ddl-15-concepts_with_counts.jinja.sql
@@ -12,7 +12,7 @@ CREATE TABLE IF NOT EXISTS {{schema}}concepts_with_counts{{optional_suffix}} AS 
             invalid_reason,
             COUNT(DISTINCT domain) AS domain_cnt,
             array_to_string(array_agg(domain), ',') AS domain,
-            SUM(total_cnt) AS total_cnt,
+            SUM(total_cnt)::bigint AS total_cnt,
             array_to_string(array_agg(distinct_person_cnt), ',') AS distinct_person_cnt
     FROM {{schema}}concepts_with_counts_ungrouped
     GROUP BY 1,2,3,4,5,6,7,8
